+++ Pandora/Paradigm/Basis/Predicate.hs
@@ -1,10 +1,14 @@
 module Pandora.Paradigm.Basis.Predicate (Predicate (..)) where
 
-import Pandora.Core.Morphism ((.), ($))
+import Pandora.Core.Morphism ((.), ($), (!))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
-import Pandora.Pattern.Object.Setoid (Boolean)
+import Pandora.Pattern.Functor.Determinable (Determinable (determine))
+import Pandora.Pattern.Object.Setoid (Boolean (True))
 
 newtype Predicate a = Predicate { predicate :: a -> Boolean }
 
 instance Contravariant Predicate where
 	f >$< g = Predicate $ predicate g . f
+
+instance Determinable Predicate where
+	determine = Predicate (True !)