+++ Pandora/Pattern/Functor.hs
@@ -9,6 +9,7 @@ import Pandora.Pattern.Functor.Extendable as Exports
 import Pandora.Pattern.Functor.Bindable as Exports
 import Pandora.Pattern.Functor.Distributive as Exports
 import Pandora.Pattern.Functor.Traversable as Exports
+import Pandora.Pattern.Functor.Determinable as Exports
 import Pandora.Pattern.Functor.Extractable as Exports
 import Pandora.Pattern.Functor.Pointable as Exports
 import Pandora.Pattern.Functor.Avoidable as Exports