+++ Pandora/Pattern/Functor/Determinable.hs
@@ -1,0 +1,6 @@
+module Pandora.Pattern.Functor.Determinable (Determinable (..)) where
+
+import Pandora.Pattern.Functor.Contravariant (Contravariant)
+
+class Contravariant t => Determinable t where
+	determine :: t a